bug bash 2

+++ src/rotors.h
@@ -16,6 +16,6 @@ void setMainPWM(uint32_t ui32Freq, uint32_t ui32Duty);
 
 void setTailPWM(uint32_t ui32Freq, uint32_t ui32Duty);
 
-extern uint32_t g_main_duty;
-extern uint32_t g_tail_duty;
+extern uint16_t g_main_duty;
+extern uint16_t g_tail_duty;
 #endif // __ROTORS_H__
